rss-agent/requirements.txt
feedparser>=6.0.0
requests>=2.28.0
PyYAML>=6.0
python-dateutil>=2.8.0
openai>=1.0.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
